+++ lib/components/layouts/community/index.tsx
@@ -1,26 +1,28 @@
 import React from 'react';
-
-import { GetSpaceDocument } from '$lib/graphql/generated/backend/graphql';
-import { isObjectId } from '$lib/utils/helpers';
-import { getClient } from '$lib/graphql/request';
-import { Space } from '$lib/graphql/generated/backend/graphql';
 import { notFound } from 'next/navigation';
+
+import { isObjectId } from '$lib/utils/helpers';
+import { getSpace } from '$lib/utils/getSpace';
+
 import { CommunityThemeProvider } from '$lib/components/features/theme-builder/provider';
 import { CommunityContainer } from './container';
 
 type LayoutProps = {
   children: React.ReactElement;
-  params: { uid: string };
+  params: { uid: string; domain: string };
 };
 
 export default async function CommunityLayout({ children, params }: LayoutProps) {
-  const { uid } = await params;
-  const variables = isObjectId(uid) ? { id: uid, slug: uid } : { slug: uid };
+  const { uid, domain } = await params;
+  let variables = {};
 
-  const client = getClient();
+  if (uid) {
+    variables = isObjectId(uid) ? { id: uid, slug: uid } : { slug: uid };
+  } else {
+    if (domain) variables = { hostname: decodeURIComponent(domain) };
+  }
 
-  const { data } = await client.query({ query: GetSpaceDocument, variables });
-  const space = data?.getSpace as Space;
+  const space = await getSpace(variables);
 
   if (!space) {
     return notFound();